VMware: Reduce volume creation time
The volumes (virtual disks) are created on datastores connected to
ESX hosts in vCenter clusters. Currently we create a list of hosts,
filter out hosts in maintenance mode (using one vCenter API call per
host), and then iterate over the filtered list to select a datastore
which is mounted to maximum number of ESX hosts. Ties are broken
based on space utilization.

This approach is not scalable for vCenter installations with large
number of ESX hosts. Also, we always consider the ESX hosts in the
same order and may end up selecting the same set of datastores. This
patch fixes these issues with the datastore selection. It creates a
list of all datastores in vCenter, which is then filtered to remove
unusable datastores and datastores not satisfying the volume requir-
ements. Finally, we select a datastore from the filtered list based
on the number of ESX hosts to which it is mounted and space utiliza-
tion.

Closes-bug: #1556902
Related-bug: #1521894
Change-Id: I0c505cbe1c82b9cbd2918a223a53a800a9bc7931