Fix quota double-spending with uniqueness-constraint
We introduce a uniqueness-constraint to avoid duplicate entries in the
quota_usages table. For introducing this, we have to delete the
duplicates already present and the soft-deleted reservations referencing
duplicate entries.

Things learned here: Do not use _and_ statement, as it will replace it
with an expression-object. Also we first tried a rank query, but now use
a join.