Change the minimal virtual disk capacity to 4MB
With our VASA provider we cannot create disks with the default minimum
of 1MB as no LUNs below 4MB size can be created. The need for creating
such small volumes comes from the fact, that finding the actual size of
stream-optimized images is left to the vCenter, which grows the volume
dynamically.