[SAP] Remove all NICs from snapshot
Until now, when creating a snapshot from an attached volume, only the
additional unwanted disks got remove and the NICs/VIFs were kept. With
nsx-t this is a problem, as it does not allow duplicate MAC addresses
on the same logical switch. Therefore, we now remove all VIFs - or
rather every device that has a "macAddress" attribute - in the
clone-process.

Change-Id: Ia7b49f09728decb6f45e2dd723b1a2cadf98279d