[SAP] Backup: send data needed to build an ImportVApp spec
This is addressing the fix for restoring VVol volumes from swift
backup. cinder-backup needs to upload the backup data to VMWare
via HttpNfc API by performing a ImportVApp call. Since this
operation is about to replace the existing backing, we want to
keep the main logic in cinder-volume. Thus, we instruct
cinder-backup how to build the spec for ImportVApp via a json-like
syntax, since it seems that suds objects can't pe pickled or
simply can't be sent plain over RPC.