[SAP] Backup: send data needed to build an ImportVApp spec
This is addressing the fix for restoring VVol volumes from swift
backup. cinder-backup needs to upload the backup data to VMWare
via HttpNfc API by performing a ImportVApp call. Since this
operation is about to replace the existing backing, we want to
keep the main logic in cinder-volume. Thus, we instruct
cinder-backup how to build the spec for ImportVApp via a json-like
syntax, since it seems that suds objects can't pe pickled or
simply can't be sent plain over RPC.

[SAP] Normalize the new backing during terminate_connection
After a restore backup operation, we correct the name and backing uuid
for the newly created backing.
This commit also includes moving some values into constants and updating
the unit tests.

[SAP] fix VM related constants
[SAP] vmdk driver - adding comment for terminate_connection()
[SAP] fix pep8

[SAP] fix volume migration for vmware
This patch adds the import_data options during volume attach
for migrating a volume.   When a volume is attached locally
for work during migration the volume needs to be writeable
in order for cinder to copy bits into the volume.   This import_data
section of the connection_properties, instructs os-brick to create
a write handle for the http connection to the volume.
This is needed for migrating a volume from one shard to another,
since cinder's generic volume copy takes over.

[SAP] Fix when migration status is None
This patch fixes an issue that tempest found during
test_volume_snapshot_backup and the volume['migration_status'] was None
and can't be dereferenced as a string.

[SAP] fix force detach when connector is None
This patch ensures that the connector is not None during
force detach.  This issue was found during a tempest run.

Change-Id: I72ce017064409d31fa940daf7597531d5dadd05c