[SAP] scheduler: Add ShardFilter
This filters out backends based on the shard defined in the backend's
extra_capabilities and in the project's tags.

[SAP] run `tox -e genopts` for "scheduler: Add ShardFilter"

Got notified about this by `tox -e pep8` ... didn't know.

[SAP] scheduler: ShardFilter lets snapshots pass

Snapshots are only run through the scheduler to check the capacity is
still available. The host is already defined and only that host is
checked. Therefore, we can savely ignore snapshots in the `ShardFilter`.

[SAP] scheduler: Allow shard override via hint

Since it would be too much effort to change our blackbox tests to use
multiple projects so they can test in all shards, we implement an
override in the `ShardFilter` via scheduler_hints.
Example:
  os volume create --size 10 asdf --hint vcenter-shard=vc-a-1

[SAP] Handle sharding-enabled in scheduler shard filter

If the project tags from keystone contain the tag "sharding_enabled"
then the backends in _all_ shards will pass the shard filter for this
project.

This was done to facilitate both enabling sharding (only one simple
tag to set), and mainly for frontend code to detect sharding status
(mostly) without parsing tag strings. (If sharding is not enabled,
then vc-* tags will have to be parsed to find out which shard(s) the
project is on.)

[SAP] ShardFilter passes everything for find_backend_for_connector

When we try to find a backend for a connector, we got called by an admin
to migrate a volume towards a connector - a specific vCenter housing a
specific VM. In those cases, we don't want to prohibit the migration
just because the volume's project doesn't have the appropriate shards
set, as the shards should more steer where we put volumes initially, not
prohibit the user from attaching a volume if it's necessary on a "stray"
instance.

Change-Id: Iccf5f5287eda536adafe925c76054f14827c467e

[SAP] Fix the shard filter for group creation

Temptest tests found a bug in the shard filter where creating
a cinder group from an existing group or group snapshot always failed
with not finding any hosts.  This is due to the shard filter not getting
much of any information about the request.

This patch adds the project_id of the group being created to ensure that
the shard filter can work properly.

[SAP] Update shard filter to filter only on vmware

This patch adds a simple check against the pool in the shard filter
to only filter if the pool is reporting as a vendor_name of 'VMware',
which is true only for vmdk.py and fcd.py drivers.

If we deploy a netapp cinder driver, this will pass the shard filter
as there is no reason to shard.

[SAP] Shard filter retype on same shard

This patch adds a check in the shard filter for volume retypes.
volume retype should always land on the same shard that the volume
already exists on.

Change-Id: I1eb7ffa179b5e5e6b2c3edee3d06393170a85cf4