[SAP] vmware: Optionally storage-profile on attach
When we set a datastore in maintenace, we remove the tag connecting it
to the storage-profile. We do this to prohibit cinder from using that
datastore for new volumes. But since cinder also checks the tags for
finding a valid datastore on attachment, it does a costly and slow
vMotion of the volume to another datasore. We don't want it to vMotion
the volumes automatically, but rather want to do it on our own, as doing
it on attachment makes attachments really slow.

Setting `vmware_profile_check_on_attach` to `false` will disable the
check on attachment. It's still done on resize, though.