[SAP] add support for reporting pools
This patch adds the reporting of each datastore that is found
in the list of storage_profiles configured, as an individual
pools.  This will expose each datastore in the profiles as
schedulable via volume types, and enable thin provisioning
limitations against each individual datastore.

Added update_provider_info call to vmware driver.

This patch modifies the volume manager's _sync_provider_info
to allow it to update the host entries for volumes.
The vmware driver adds an update_provider_info() call
and loops through all the volumes and updates the host entry
to include the pool name at the end of the entry.   This is
all done at cinder-volume service startup, which enables every
volume existing in the cinder db to get an updated host entry,
so after the pools are reported, cinder can still manage
the existing volumes properly for api calls.
If the host entry isn't updated, then Cinder will not be able to
correctly take action on the volume as the host won't be running.

[SAP] Update stats for each storage pool

This patch updates the volume stats reporting for the vmdk
driver to include datastores as pools when it's enabled and
ensures no pools are reported when disabled.

SAP Fix extra capabilities with pools.

This patch ensures that extra_capabilities are inserted into
each pool capabilities for the driver.

Ensure migrate works with pools

This patch updates the remote and select_ds_for_volumes to allow
passing in a specific cinder_host, which happens during volume
migration.

Ensure we handle no backing volumes

this patch ensures that we handle volumes that are still assigned to
a backend_name for the pool during attach time.  This is because
we have lazy creates enabled and some volumes don't have any
backing at the time of attach.  We try and fallback to the old
mechanism of assigning a datastore/pool

[SAP] Update connection_capabilities reporting in stats

This patch fixes an issue with reporting the connection capabilities
in pools enabled stats for vmdk driver.

[SAP] Ensure pools specify thick_provisioning_support enabled

This patch forces the pool stats to say they support
thick provisioning.

[SAP] Remove the tracing from _get_datastores_for_profiles

This tracing was useful during development of the datastores
as pool, now it's just noise.

[SAP] Add any custom attributes to pool stats

This patch tries to read any custom attributes associated with
each datastore and shove those attributes in the pool stats.
This will allow us to filter by custom attributes.

The end goal is to add a netapp backend fqdn custom attribute
so we know where each datastore lives.  This will allow cinder
to have affinity/anti-affinity to netapps.

[SAP] Added pool_state tracking. Updated backend_state

This patch adds the ability to track if a datastore is available
to be used in get_volume_stats.  If a datastore is in maintenance
mode, then it can't accept volume provisioning requests.  So the
driver now marks the datastore/pool as 'down'.  When the datastore
comes back out of maintenance mode, then the pool_state is changed to
'up'.

Currently, Cinder doesn't have a built in mechanism for dealing with a
pool_state being down, but we can simulate it by adding an extra spec
setting in all volume types of pool_state='up'.  That way the
capability filter will filter out all pools that don't have that
capability coming from the driver.

This patch also changes where the backend_state is set so that
the scheduler can mark the service as down correctly.

[SAP] Fix multiattach reporting for vmdk

This patch fixes the reporting of multiattach capability in the
vmdk driver.   Multiattach -> multiattach

[SAP] change storage_profile reporting

This patch changes the way the storage_profile attribute is reported
in the pool stats.  Previously the value was a dictionary that
included the name and the storage profile id.   The storage_profile
needs to only be the name so we can match against it in the
volume type extra spec.

This is related to the cinder-issue:
https://github.wdf.sap.corp/cc/cinder-issues/issues/14

Change-Id: I3a5584d858ca198edb48c8ea482924c1a1b0ee69