[SAP] add support for reporting pools
This patch adds the reporting of each datastore that is found
in the list of storage_profiles configured, as an individual
pools.  This will expose each datastore in the profiles as
schedulable via volume types, and enable thin provisioning
limitations against each individual datastore.

Added update_provider_info call to vmware driver.

This patch modifies the volume manager's _sync_provider_info
to allow it to update the host entries for volumes.
The vmware driver adds an update_provider_info() call
and loops through all the volumes and updates the host entry
to include the pool name at the end of the entry.   This is
all done at cinder-volume service startup, which enables every
volume existing in the cinder db to get an updated host entry,
so after the pools are reported, cinder can still manage
the existing volumes properly for api calls.
If the host entry isn't updated, then Cinder will not be able to
correctly take action on the volume as the host won't be running.

[SAP] Update stats for each storage pool

This patch updates the volume stats reporting for the vmdk
driver to include datastores as pools when it's enabled and
ensures no pools are reported when disabled.

SAP Fix extra capabilities with pools.

This patch ensures that extra_capabilities are inserted into
each pool capabilities for the driver.

Ensure migrate works with pools

This patch updates the remote and select_ds_for_volumes to allow
passing in a specific cinder_host, which happens during volume
migration.

Ensure we handle no backing volumes

this patch ensures that we handle volumes that are still assigned to
a backend_name for the pool during attach time.  This is because
we have lazy creates enabled and some volumes don't have any
backing at the time of attach.  We try and fallback to the old
mechanism of assigning a datastore/pool