[SAP] Chunkeddriver - no seek (#30)
* chunkeddriver - improving the restore operation

Compute all the incremental backups prior to writing it to the
file, so that a single write operation is executed on the
volume_file regardless the number of incremental backups.
Removes the need of seeking back into the volume_file for
overwriting with incremental chunks.

* Fix add_object() to avoid unnecessary iteration over the same object

The previous approach of iterating on enumerate() while inserting
2 times into the list, was doing an extra useless iteration over
an object that has just been inserted.
We switch to while loop so that we are able to jump to the desired
index after we inserted the segments into the list.

* Add iterator methods in BackupRestoreHandle

Since this was built to be used as an iterator, it's cleaner
to use the python iterator api and get rid of the has_next()
and get_next() methods.

* Fix _clear_reader() to properly clear the reader if it's not needed

It checks if there are no more segments of the same object after the
current index till the end of the segments list, case when it also
closes and removes the reader from the cache directly.

* Added a docstring for the Segment.of() method

* Create BackupRestoreHandleV1 for handling v1 metadata

Since we're handling most of the restore process within the
BackupRestoreHandle class, we're now moving the metadata
versioning down to it's own class (BackupRestoreHandleV1).

DRIVER_VERSION_MAPPING should now refer to class names.
This kind of classes should extend BackupRestoreHandle
or at least take as constructor parameters:
* chunked_driver - an instance of ChunkedBackupDriver
* volume_id - the volume id
* volume_file - the file handle where to write the data

Additionaly, such a class should implement the following methods:
* add_backup(backup, metadata) - called for each backup
* finish_restore() - called after the backups are iterated

* Make BackupRestoreHandle an abstract class

Since BackupRestoreHandle does not implement the add_backup method
which lets other classes inheriting it to define their own backup
and metadata handling, it makes sense to make it abstract.

[SAP] Improve the deletion of object readers in BackupRestoreHandle (#51)

* Improve the deletion of object readers in BackupRestoreHandle

This deletes right away an object reader which is not needed
anymore, without waiting for the garbage collector to take care
of it. As an effect, this should lower the memory consumption.

* Fix finish_restore to increment the current object index

This fixes the high memory consumption caused by _clear_reader
which was not able to do its job because the self._idx
was not reflecting the correct value.

* fix pep8 - use delayed string interpolation

[SAP] Fix pep8, missing backup restore feature

This patch updates the SAP modified chunkedbackup driver
to include raising the exception when a user cancels the
restore process.  That fixes a failure in the unit tests for
the nfs driver expected exception raising of BackupRestoreCancel.

This patch also fixes some pep8 issues with utils being renamed
volume_utils in train.

[SAP] set the correct vmdk_size when restoring backups

VMDKs are stream optimized, having different size than the actual
volume. We have to set the correct vmdk_size to VmdkWriteHandle
so that it can validate the upload progress correctly.

Fix setting the restore handle size

This patch changes how cinder calls the oslo.vmware rw_handle
to set the size of the contents of the vmdk file bytes to using
a method on the object.

This requires PR:
https://github.com/sapcc/oslo.vmware/pull/35

Change-Id: I2fb08b0756d47a202f55cc5c5faee25398c4eb74