[SAP] add concourse_unit_test_task
This patch adds the concourse_unit_test_task file so we can run
unit tests during the loci image build process.

[SAP] upgrade pip for concourse_unit_test_task

There is a bug in setuptools that prevents python installs
from working correctly and you end up with an error
"error: 'egg_base' must be a directory name (got `src`)"

This patch upgrades the version of pip for running unit tests,
which should fix the error.

[SAP] Add libpq-dev to concourse_unit_test_task

Change-Id: I120638f993eba0ffdfe68ea6c8c2ec3d4ba3069e

This patch updates the running of the concourse unit test task
against train to include the libpq-dev package as well as make
sure we run py3 tox.

[SAP] fix concourse_unit_test_task for 2023.1

Change-Id: I6089ba72f789d50e3e5ec69280cceb76813fad37