SAP: fix volume migration for vmware
This patch adds the import_data options during volume attach
for migrating a volume.   When a volume is attached locally
for work during migration the volume needs to be writeable
in order for cinder to copy bits into the volume.   This import_data
section of the connection_properties, instructs os-brick to create
a write handle for the http connection to the volume.

This is needed for migrating a volume from one shard to another,
since cinder's generic volume copy takes over.