[SAP] Allow a driver to specify additional RPC endpoints
Drivers may need to create their own RPC communication channel to
exchange vendor-specific logic.
A driver can now specify additional RPC endpoints by appending it
to the `self.additional_endpoints` list. The manager will always
initialize that property as an empty list and take care of
exposing the endpoints after the driver has been loaded.
This is a feature which is already possible in Nova.