[SAP] implementation of volume migration
Implement driver's `migrate_volume`.
It adds a new RPC server and client for communicating with the
destination host to get information needed to build the
relocate spec for a particular volume: datastore, resource pool,
host, vcenter info, and to perform `move_backing_to_folder` after
the volume was relocated, or, to create the backing on the
destination host if it doesn't already exist.

[SAP] Reporting location_info

Report location_info in the capabilities.
It is a concatenation of a static name and the current vcenter
service instanceUuid.
The location_info is being used in the migrate_volume
to make sure we don't attempt to migrate non-vmware volumes.

Additionally, this commit removes the creation of the backing
on the dest_host, in case of a migration for a volume which does
not already have a backing created. The backing will be created
on initialize_connection anyway.

Update vmdk.py- Fix ssl_thumbprint

Fix ssl_thumbprint, the type expected is utf-8 string instead of byte array

Change-Id: Ia2e9de8614ce217aecffb74601cca8290ec63fcc