[SAP] implementation of volume migration
Implement driver's `migrate_volume`.
It adds a new RPC server and client for communicating with the
destination host to get information needed to build the
relocate spec for a particular volume: datastore, resource pool,
host, vcenter info, and to perform `move_backing_to_folder` after
the volume was relocated, or, to create the backing on the
destination host if it doesn't already exist.