[SAP] Reporting location_info
Report location_info in the capabilities.
It is a concatenation of a static name and the current vcenter
service instanceUuid.
The location_info is being used in the migrate_volume
to make sure we don't attempt to migrate non-vmware volumes.

Additionally, this commit removes the creation of the backing
on the dest_host, in case of a migration for a volume which does
not already have a backing created. The backing will be created
on initialize_connection anyway.