Ability to reschedule and migrate volumes upon attachment
Given a volume which is being scheduled to locaion A
and a connector (VM) which resides in location B and wants to
attach that volume. The volume backend is now able to reject
that connector (for example if A is not reachable by B) by
throwing an exception.
In this case, cinder will reschedule the volume to a new host
filtering by the `connection_capabilities` provided by the volume
backend and by the connector.
All `connection_capabilities` provided by the connector must be
in the `connection_capabilities` exposed by the volume backend for
this scheduling to take place.
Once a new host is selected, a migration is performed and
eventually the connection_info is passed back to the connector.

This functionality is disabled by default and can be enabled by
setting `allow_migration_on_attach=True`.

This commit also includes an implementation for VMwareVcVmdkDriver,
which is used to reject connectors comming from different vCenter
services.