[SAP] Improve the deletion of object readers in BackupRestoreHandle (#51)
* Improve the deletion of object readers in BackupRestoreHandle

This deletes right away an object reader which is not needed
anymore, without waiting for the garbage collector to take care
of it. As an effect, this should lower the memory consumption.

* Fix finish_restore to increment the current object index

This fixes the high memory consumption caused by _clear_reader
which was not able to do its job because the self._idx
was not reflecting the correct value.

* fix pep8 - use delayed string interpolation