[SAP] Create volume from image with expected size
When we're creating a volume from an image, the size was set to 0 to let
the vCenter figure out the uncompressed size. But with our VASA-provider
and using thin-provisioning we end up with empty volumes, presumably
because the volume doesn't auto-grow properly.

Since we know the size the volume should have in the end in advance and
we're not re-using the created VM as a template for others, we can
create the volume initially with the user-wanted size and don't have to
resize the volume afterwards.