[SAP] Randomize the selection of best datastores
This patch adds a new config option that allows for randomizing
the selection of datastores at backing creation time.  First the
best datastores are chosen, which are the most connected datatstores
with the most space available.  Then that list of datastores is
randomized for selection.  The first datastore in that random list is
chosen.  This helps the driver not always pick the same datastore
when each datastore reports lots of free space.

Change-Id: I3992fa1a0cb5f059140685c74f70f8047bbb1480

[SAP] Fix randomize_datastore selection.

This patch passes in the randomization settings to the datastore
selector object during driver do_setup() time.  Previously
those values were passed ony if the ds object didn't already exist.