[SAP] Randomize the selection of best datastores
This patch adds a new config option that allows for randomizing
the selection of datastores at backing creation time.  First the
best datastores are chosen, which are the most connected datatstores
with the most space available.  Then that list of datastores is
randomized for selection.  The first datastore in that random list is
chosen.  This helps the driver not always pick the same datastore
when each datastore reports lots of free space.

Change-Id: I3992fa1a0cb5f059140685c74f70f8047bbb1480