[SAP] Fix pep8, missing backup restore feature
This patch updates the SAP modified chunkedbackup driver
to include raising the exception when a user cancels the
restore process.  That fixes a failure in the unit tests for
the nfs driver expected exception raising of BackupRestoreCancel.

This patch also fixes some pep8 issues with utils being renamed
volume_utils in train.