[SAP] fix DB migration scripts
This patch updates the migrate_versions scripts for the upgrade
from queens all the way to Train.   The upstream Train release
collapsed all of the migrate scripts into a single script
called 123_cinder_init.py.  Since we are upgrading from queens
we had to add placeholders for version 120, 121, 122, and also
replace the 123_cinder_init.py with the rocky branch script for
version 123, which is 123_add_transfer_no_snapshots