[SAP] Add db 120_cinder_init.py migration
Upstream cinder train refactored all of the previous db migration
scripts into 123_cinder_init.py to collapse all of the migrations
into a single file.  In order to keep the unit tests working, I had
to collapse the db migration scripts into 120_cinder_init.py which
includes our db patches, as our db schema version in queens ended
with version 120.  This ensures that 120 will be skipped when we
upgrade from queens to train, and the first upgrade script that runs
will start with 121.

This patch fixes an issue of tox -epy27 failing in test.