Set purge_props=False by default in GlanceImageService update
The glance service update() method removes by default all the
existing properties and overwrites with the ones passed as parameter.
This is against what calling services expect to [1] and opposite to
how the image service is implemented in nova [2].
Therefore, the expectation is that the update() actually updates the
image properties dict. Purging the image properties that are not
specified in the input parameter should be optional and should be
controlled by the caller when it knows what it's doing.

[1] https://github.com/openstack/oslo.vmware/blob/543ef8ba36295d7079c594adb80d252f9e689d83/oslo_vmware/image_transfer.py#L327
[2] https://github.com/openstack/nova/blob/1516fdd3944dc8bb655664c136bb16715b0c1df2/nova/image/glance.py#L531