[SAP] Force volume create from backup to restoring
This patch fixes an issue with calling create volume from backup.
There are 2 ways to create a volume from a backup.
1) calling cinder backup-restore
2) calling cinder create --backup-id

There is a series of issues with using #2 with vmware due to:
A) the volume is put into 'creating' status before calling the
   backup manager to restore the bits to the new raw volume.
B) the vmware driver uses the volume status field to determine if
   it should add required information to be returned in
   initialize_connection() to exist for os-brick to return a
   VmdkWriteHandle vs a VmdkReadHandle.

Both A and B combined results in 100% failure rate for creating a
volume from back up using technique #2.

Changing the volume status to restoring after the raw cinder volume
has been created fixes this issue.

Change-Id: Ibf6dd8d20b1bbe0c745e517f7617fb3ecf7583f1