[SAP] Create empty backing for live-migration
The actual live-migration is happening in nova,
but nova needs to know where to place the volume.

An empty backing vm will be created, and nova takes care of the rest.
The `datastore` in the connection info will be a mo-ref specifying the
target destination (or initial location) and won't get updated over time.

Change-Id: Iae01956daf0b711e0170c034f8b0597129e337ac