[SAP] WIP call correct host for live migration
This patch alters the cinder volume api detach call to check the
attachment info of a volume to ensure that the
connection_info['data']['volume_host'] is the same as the
volume.service_topic_queue.  If it's not, then ensure the entry in
the connection_info is the target of the rpc call.