Cleanup glance image metadata during instance reschedule
When we create an instance using volume, if the instance
create fails the first time, it tries to reschedule it using
the same volume.
When rescheduling, we encounter an issue with improper cleanup
of glance image metadata, hence it fails on cinder side.
This patch addresses it by checking for glance image metadata
incase of an exception and does the proper cleanup.

Closes-Bug: #1936474

Change-Id: Ife4e330b11bfe278231aa93e354e12b13d7dd7fc