[SAP] Filter out hosts that are marked buildup
This patch filters out hosts in a cluster that has been
marked by ops as 'buildup'.  This ensures that we don't use
hosts that aren't ready to be used quite yet.

Ops has to set a 'buildup' : 'true' custom attribute in the
cluster.  This patch will reverse lookup which cluster a host is in
and ensure that the cluster isn't marked as buildup.

Change-Id: If2fcbe31dd2ed50dcc8f72d2c573761f5d8b2058