[SAP] Add vmware extension for migration
This is a POC/WIP to add the extension to the cinder API
to allow calling os-migrate_volume_by_connector.  This will
still be an admin API endpoint as it will call migrate_volume
underneath it.

Notes: I have tested this extension on my home devstack and it worked.

[SAP] Fixed name of migrate_volume method name

This fixes a conflict of the method name for the _migrate_volume
(by connector) in the vmware_migrate extension.

[SAP] Change the API return code for ConnectorRejected

This patch changes the HTTP response code for a call to
attach a volume where the ConnectorRejected is raised.
The Response code was 500 generic error.  The new
reponse is HTTP 406 HTTPNotAcceptable

[SAP] put volume in maintenance mode during migrate

This patch ensures that the os migrate by connector
puts the volume in maintenance mode if the lock_volume
flag is passed in the API to migrate by connector

Change-Id: If8b7f4cfafc5b94878ae985d6081e94b8d62f522

[SAP] Fix log.debug entry for migrate_by_connector

Small issue with the logging text string format.

[SAP] Handle reserved status in live-migrations

As we now lock the volumes, the statu of the volume
will change from in-use to reserved.
We have to handle that state, and look instead at the
attach_status to determine if the volume is attached

[SAP] Fix migration by connector filter_properties

This patch fixes the migration by connector with affinity/anti-affinity
filter_properties.  The previous code was doing an rpcapi call to the
volume manager to migrate the volume with the filter properties.
At this point the scheduler has already picked a destination host, so
the filter_properties being passed to the volume manager is useless.

This patch sends the filter_properties through the scheduler to help it
pick the right backend before calling the volume manager to migrate
the volume to the new backend.

Change-Id: Id3a6442c76d40c0dfc1613f887fade3af026506e