[SAP] add SAPLargeVolumeFilter
This filter disallows any volumes larger than 2TB to be created on
any datastore that has 'vvol' in it's name.

[SAP] Fix an issue with SAPLargeVolumeFilter

The filter was getting a KeyError during snapshot_create
requests.

This patch reworks the filter to use the safety like the capacity
filter uses for getting the size out of the request spec.

[SAP] update size check to Gib units

This patch updates the 2 T units check to 2048, for Tib vs Tb

Change-Id: I20c2ef56474af7163ac4bcaa6a0f86f0f191b0c6