[SAP] Account for volumes already provisioned to a host
This patch updates the volume manager's creation of the
allocated_capacity_gb setting for valid volumes that have
already been provisioned.  The volume manager looks at all
non-deleted volumes in the database to account for storage
that has been allocated by cinder in the backend.
This patch updates that check for volumes that are in other
non error states.