Vmware: Remove nvp.vm-uuid extraConfig
The Nova vmwareapi driver falls back to the extraConfig
option value nvp.vm-uuid to identify vms, if it cannot
find VMs by the config.instanceUuid attribue.
A clone copies the attribute, so when nova retries
to delete a vm after initially succeeding,
it will find the backing with the same value instead
and delete it instead.

Change-Id: Iaaf12e5e43ef8465d2b0c4bb059deb2acc379c98