[SAP] Rework Capacity filter
This patch adds the new cinder/utils.py calculate_capacity_factors
to provide a detailed and more consistent,accurate view of the
various factors in determining virtual free space for a particular
backend.  It takes into consideration total capacity, free space,
thin/thick provisioning in the volume type, thin/thick support in the
backend, as well as reserved percentage and max_over_subscription_ratio.

Since the vmware driver is configured to allow lazy creation of volumes,
the free space reported by the pool/datastore isn't a reliable
source of how much is free considering what has been requested from
cinder to allocate.

This patch calculates what should be free based upon the
total available capacity ( total - reserved ) and what cinder
has tracked as allocated against that backend.
If that calculated free is less than the reported free, then
the calculated free is what is reported as virtual_free_space.

There is a known issue in cinder with keeping track of the allocated
space in 2 places:
1) at startup cinder only considers volumes that are in-use and
   available.  All other volumes in other states aren't used to
   calculate the allocated space.   This has to be fixed.
   This is fixed here: https://github.com/sapcc/cinder/pull/117

2) The allocated space isn't adjusted during volume migrations.

[SAP] Replace CapacityFilter

This patch reworks the CapacityFilter which uses the new
utils.calculate_capacity_factors to determine if the
virtual free space available.

The main issue is that the vmware driver does lazy creates.
This causes an issue of over reporting of free space by
the backend, because the space hasn't been consumed yet.
So the amount of free space is not accurate with respect to
how much has been allocated by cinder.  This updated calculate
capacity factors as well as the SAPCapacityFilter accounts for
the virtual free space by using the cinder allocated capacity
tracking.  If the free space is reported less than what cinder
thinks should be available, then the reported free space is used.

This relies on an acurate reporting of the allocated capacity
by the driver.  We know there is an issue with allocated capacity
not being reported correctly for migrated volumes, as well as
accounting for existing volumes at startup.  The startup issue
should be solved with this PR:
https://github.com/sapcc/cinder/pull/117

Will have to do a folllow up to account for updating the allocated
capacity for migrated volumes.

Change-Id: I26bf0cb589e8295556b625978b38a1d4c17f3f34